requirements.txt
# Core dependencies
openai>=1.0.0
google-generativeai>=0.8.0
requests>=2.28.1
python-dotenv>=1.0.0
